+++ src/main/java/info3/game/RemoteView.java
@@ -17,7 +17,7 @@ public class RemoteView extends View {
 	}
 
 	@Override
-	public Avatar createAvatar(int id, Vec2 pos, String filename, int imageLen, int animationSpeed) {
+	public Avatar createAvatar(int id, Vec2 pos, String filename, int imageLen, long animationSpeed) {
 		RemoteAvatar av = new RemoteAvatar(id, (LocalController) this.controller, filename, imageLen, animationSpeed);
 		av.setPosition(pos);
 		this.avatars.put(id, av);
